Rename common to utils
