+++ lib/ts/utils/cookieHandler/defaultImplementation.ts
@@ -14,7 +14,7 @@
  */
 import { CookieHandlerInterface } from "./types";
 
-function getWindowOrThrow(): Window {
+export function getWindowOrThrow(): Window {
     if (typeof window === "undefined") {
         throw Error(
             "If you are using this package with server-side rendering, please make sure that you are checking if the window object is defined."
