+++ lib/ts/utils/cookieHandler/types.ts
