+++ lib/ts/utils/windowHandler/defaultImplementation.ts
@@ -14,7 +14,7 @@
  */
 import { WindowHandlerInterface } from "./types";
 
-function getWindowOrThrow(): Window {
+export function getWindowOrThrow(): Window {
     if (typeof window === "undefined") {
         throw Error(
             "If you are using this package with server-side rendering, please make sure that you are checking if the window object is defined."
