+++ lib/ts/utils/windowHandler/index.ts
