+++ lib/ts/utils/windowHandler/types.ts
