+++ utils/cookieHandler/types.js
@@ -17,4 +17,4 @@ function __export(m) {
     for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
 }
 exports.__esModule = true;
-__export(require("../../lib/build/common/cookieHandler/types"));
+__export(require("../../lib/build/utils/cookieHandler/types"));
